requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
